Added a new resource
Hi, I have added a new resource for the C++ function in the document. I found this tutorial online while looking for some resources to learn c++. Topics were explained pretty well and found it to be useful and thought of adding it here. Hope this small contribution will help other learners.